Сетевые функции вынесены в отдельный класс
