+++ VKInterface/Controllers/FriendsViewContoroller/FriendsViewController.swift
@@ -8,6 +8,7 @@
 import UIKit
 import Alamofire
 
+
 class FriendsViewController: UIViewController {
   
   @IBOutlet weak var tableView: UITableView!
@@ -18,30 +19,10 @@ class FriendsViewController: UIViewController {
   let cellReuseIdentificator = "cellReuseIdentificator"
   let toGallerySeague = "toGallerySeague"
   var friendsArray = [Friends]()
+  var usersArray: [UsersListResponse] = []
+  private let networkServices = NetworkServices()
   
-  var usersArray = [UsersListResponse]()
   
-  private func vkFriendsList(_ completion: @escaping (UsersListResponse) -> Void ) {
-    
-    let userId = Session.shared.userId
-    let token = Session.shared.token
-    let host = "https://api.vk.com"
-    let path = "/method/friends.get"
-    
-    let url = URL(string: host + path)!
-    let parameters = [
-      "access_token" : token,
-      "owner_id" : userId,
-      "order" : "hints",
-      "fields" : "photo_200_orig",
-      "name_case" : "nom",
-      "v" : "5.131"
-    ]
-    AF.request(url, parameters: parameters).responseDecodable{ (response: AFDataResponse<UsersListResponse>)  in
-      guard let response = response.value else { return }
-      completion(response)
-    } .resume()
-  }
   
   ////    let friendsData = FriendsData.shared.sourceFriendsArray
   
@@ -52,10 +33,8 @@ class FriendsViewController: UIViewController {
     tableView.delegate = self
     FriendsData.shared.delegate = self
     
-    vkFriendsList()
-    { response in
-      
-      self.usersArray.append(response)
+    networkServices.vkFriendsList()  { response in
+      self.usersArray = [response]
       print("Финальный массив \(self.usersArray)")
     }
     
@@ -65,8 +44,8 @@ class FriendsViewController: UIViewController {
                                            selector: #selector(addNewUser(_ :)),
                                            name: Notification.Name("addNewUserButton"),
                                            object: nil)
-        FriendsData.shared.fillFriendsData()
-        friendsArray = FriendsData.shared.sourceFriendsArray
+    FriendsData.shared.fillFriendsData()
+    friendsArray = FriendsData.shared.sourceFriendsArray
     tableView.reloadData()
   }
