+++ VKInterface/NetworkServices/vkprintDataToConsole.swift
